+++ src/BodyPlugin/IoConnectionView.cpp
@@ -531,7 +531,7 @@ IoConnectionView::Impl::Impl(IoConnectionView* self)
 
     // Setup the table
     auto hframe = new QFrame;
-    hframe->setFrameStyle(QFrame::HLine | QFrame::Sunken);
+    hframe->setFrameStyle(static_cast<int>(QFrame::HLine) | static_cast<int>(QFrame::Sunken));
     vbox->addWidget(hframe);
     setFrameShape(QFrame::NoFrame);
     setSelectionBehavior(QAbstractItemView::SelectRows);
